+++ ext/opentelemetry-ext-pyramid/CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## Unreleased
 
+- Use one general exclude list instead of two ([#872](https://github.com/open-telemetry/opentelemetry-python/pull/872))
+
 ## 0.9b0
 
 Released 2020-06-10